+++ apps/admin_panel/assets/src/omg-page-token/index.js
@@ -167,6 +167,7 @@ class TokenDetailPage extends Component {
           onFetchSuccess={fetch}
         />
         <ExchangePairModal
+          action='create'
           open={this.state.createExchangePairModalOpen}
           onRequestClose={this.onRequestCloseCreateExchangePair}
         />
